Refactor the pegdown markdown rendering into a concrete class.
Additional file renaming.
